requirements.txt
python-telegram-bot==13.7
urllib3==1.26.7